ready contollers

+++ src/main/java/home/proj/BookStore/repository/AuthorRepository.java
@@ -4,6 +4,7 @@ import home.proj.BookStore.entity.Author;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
 @Repository
-public interface AuthorRepository extends JpaRepository<Author, Integer> {
+public interface AuthorRepository extends JpaRepository<Author, Long> {
 }
